// packages/datadog-api-client-v2/models/SecurityFilterExclusionFilter.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Exclusion filter for the security filter.
*/
export declare class SecurityFilterExclusionFilter {
    /**
    * Exclusion filter name.
    */
    'name': string;
    /**
    * Exclusion filter query. Logs that match this query are excluded from the security filter.
    */
    'query': string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SecurityFilterExclusionFilter;
    static serialize(data: SecurityFilterExclusionFilter): {
        [key: string]: any;
    };
    constructor();
}